fix #7342 chore(project): move notifications into legacy
Because

* Notifications are only used by legacy

This commit

* Moves notifications app into legacy
