+++ app/experimenter/legacy/bugzilla/tests/test_tasks.py
@@ -11,7 +11,7 @@ from experimenter.legacy.bugzilla.tests.mixins import MockBugzillaMixin
 from experimenter.legacy.legacy_experiments.models import Experiment
 from experimenter.legacy.legacy_experiments.tests.factories import ExperimentFactory
 from experimenter.legacy.normandy.tests.mixins import MockNormandyMixin
-from experimenter.notifications.models import Notification
+from experimenter.legacy.notifications.models import Notification
 
 
 class TestCreateBugTask(MockRequestMixin, MockBugzillaMixin, TestCase):
