+++ app/experimenter/legacy/legacy_experiments/tests/test_forms.py
@@ -36,7 +36,7 @@ from experimenter.legacy.legacy_experiments.forms import (
 )
 from experimenter.legacy.legacy_experiments.models import Experiment
 from experimenter.legacy.legacy_experiments.tests.factories import ExperimentFactory
-from experimenter.notifications.models import Notification
+from experimenter.legacy.notifications.models import Notification
 from experimenter.openidc.tests.factories import UserFactory
 from experimenter.projects.tests.factories import ProjectFactory
