+++ app/experimenter/legacy/notifications/tests/factories.py
@@ -1,7 +1,7 @@
 import factory
 from faker import Factory as FakerFactory
 
-from experimenter.notifications.models import Notification
+from experimenter.legacy.notifications.models import Notification
 from experimenter.openidc.tests.factories import UserFactory
 
 faker = FakerFactory.create()
